+++ R/ddml_late.R
@@ -55,10 +55,16 @@
 #'     Omission of the \code{args} element results in default arguments being
 #'     used in \code{fun}. Omission of \code{assign_X} (and/or \code{assign_Z})
 #'     results in inclusion of all variables in \code{X} (and/or \code{Z}).
-#' @param learners_DXZ Optional argument to allow for different estimators of
-#'     \eqn{E[D \vert X, Z]}. Setup is identical to \code{learners}.
-#' @param learners_ZX Optional argument to allow for different estimators of
-#'     \eqn{E[Z\vert X]}. Setup is identical to \code{learners}.
+#' @param learners_DXZ,learners_ZX Optional arguments to allow for different
+#'     estimators of \eqn{E[D \vert X, Z]}, \eqn{E[Z \vert X]}. Setup is
+#'     identical to \code{learners}.
+#' @param custom_ensemble_weights_DXZ,custom_ensemble_weights_ZX Optional
+#'     arguments to allow for different
+#'     custom ensemble weights for \code{learners_DXZ},\code{learners_ZX}. Setup
+#'     is identical to \code{custom_ensemble_weights}. Note:
+#'     \code{custom_ensemble_weights} and
+#'     \code{custom_ensemble_weights_DXZ},\code{custom_ensemble_weights_ZX} must
+#'     have the same number of columns.
 #' @param subsamples_Z0,subsamples_Z1 List of vectors with sample indices for
 #'     cross-fitting, corresponding to observations with \eqn{Z=0} and
 #'     \eqn{Z=1}, respectively.
@@ -138,6 +144,9 @@ ddml_late <- function(y, D, Z, X,
                       ensemble_type = "nnls",
                       shortstack = FALSE,
                       cv_folds = 5,
+                      custom_ensemble_weights = NULL,
+                      custom_ensemble_weights_DXZ = custom_ensemble_weights,
+                      custom_ensemble_weights_ZX = custom_ensemble_weights,
                       subsamples_Z0 = NULL,
                       subsamples_Z1 = NULL,
                       cv_subsamples_list_Z0 = NULL,
@@ -148,7 +157,6 @@ ddml_late <- function(y, D, Z, X,
   is_Z0 <- which(Z == 0)
   nobs_Z0 <- length(is_Z0)
   nobs_Z1 <- nobs - nobs_Z0
-  nensb <- length(ensemble_type)
 
   # Create sample fold tuple by treatment
   if (is.null(subsamples_Z0) | is.null(subsamples_Z1)) {
@@ -200,6 +208,7 @@ ddml_late <- function(y, D, Z, X,
   y_X_Z0_res <- get_CEF(y[is_Z0], X[is_Z0, , drop = F],
                         learners = learners, ensemble_type = ensemble_type,
                         shortstack = shortstack,
+                        custom_ensemble_weights = custom_ensemble_weights,
                         cv_subsamples_list = cv_subsamples_list_Z0,
                         subsamples = subsamples_Z0,
                         silent = silent, progress = "E[Y|Z=0,X]: ",
@@ -209,6 +218,7 @@ ddml_late <- function(y, D, Z, X,
   y_X_Z1_res <- get_CEF(y[-is_Z0], X[-is_Z0, , drop = F],
                         learners = learners, ensemble_type = ensemble_type,
                         shortstack = shortstack,
+                        custom_ensemble_weights = custom_ensemble_weights,
                         cv_subsamples_list = cv_subsamples_list_Z1,
                         subsamples = subsamples_Z1,
                         silent = silent, progress = "E[Y|Z=1,X]: ",
@@ -228,6 +238,7 @@ ddml_late <- function(y, D, Z, X,
                           learners = learners_DXZ,
                           ensemble_type = ensemble_type,
                           shortstack = shortstack,
+                          custom_ensemble_weights = custom_ensemble_weights_DXZ,
                           cv_subsamples_list = cv_subsamples_list_Z0,
                           subsamples = subsamples_Z0,
                           silent = silent, progress = "E[D|Z=0,X]: ",
@@ -248,6 +259,7 @@ ddml_late <- function(y, D, Z, X,
                           learners = learners_DXZ,
                           ensemble_type = ensemble_type,
                           shortstack = shortstack,
+                          custom_ensemble_weights = custom_ensemble_weights_DXZ,
                           cv_subsamples_list = cv_subsamples_list_Z1,
                           subsamples = subsamples_Z1,
                           silent = silent, progress = "E[D|Z=1,X]: ",
@@ -258,10 +270,15 @@ ddml_late <- function(y, D, Z, X,
   Z_X_res <- get_CEF(Z, X,
                      learners = learners_ZX, ensemble_type = ensemble_type,
                      shortstack = shortstack,
+                     custom_ensemble_weights = custom_ensemble_weights_ZX,
                      cv_subsamples_list = cv_subsamples_list,
                      subsamples = subsamples,
                      compute_insample_predictions = F,
                      silent = silent, progress = "E[Z|X]: ")
+
+  # Update ensemble type to account for (optional) custom weights
+  ensemble_type <- dimnames(y_X_Z0_res$weights)[[2]]
+  nensb <- ifelse(is.null(ensemble_type), 1, length(ensemble_type))
 
   # Check whether multiple ensembles are computed simultaneously
   multiple_ensembles <- nensb > 1