+++ tests/testthat/test-ddml_fpliv.R
@@ -85,6 +85,30 @@ test_that("ddml_fpliv computes with multiple ensemble procedures", {
   expect_equal(length(ddml_fpliv_fit$coef), 4)
 })#TEST_THAT
 
+test_that("ddml_fpliv computes with custom weights", {
+  # Simulate small dataset
+  nobs <- 200
+  X <- cbind(1, matrix(rnorm(nobs*39), nobs, 39))
+  Z <- matrix(rnorm(nobs*10), nobs, 10)
+  UV <- matrix(rnorm(2*nobs), nobs, 2) %*% chol(matrix(c(1, 0.7, 0.7, 1), 2, 2))
+  D <-  X %*% runif(40) + Z %*% c(1, runif(9)) + UV[, 1]
+  y <- D + X %*% runif(40) + UV[, 2]
+  # Define arguments
+  learners <- list(list(fun = ols),
+                   list(fun = ols))
+  # Compute LIE-conform DDML IV estimator
+  ddml_fpliv_fit <- ddml_fpliv(y, D, Z, X,
+                               learners,
+                               ensemble_type = c("average"),
+                               cv_folds = 3,
+                               custom_ensemble_weights = diag(1, 2),
+                               sample_folds = 3,
+                               silent = T)
+
+  # Check output with expectations
+  expect_equal(length(ddml_fpliv_fit$coef), 3)
+})#TEST_THAT
+
 test_that("ddml_fpliv computes with multiple ensembles w/o the LIE", {
   # Simulate small dataset
   nobs <- 200