+++ NEWS.md
@@ -2,6 +2,7 @@
 
 - Adds support for the average treatment effect on the treated estimator.
 - Adds support for local average treatment effect estimation with perfect compliance or perfect non-compliance.
+- Adds support for custom ensemble weights.
 
 # ddml 0.1.0